[pip] wire up existing pynacl_validator in new app
