+++ drp_aa_mvp/data_rights_request/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path
 from . import views
-from . import pynacl_validator
 
 
 urlpatterns = [
@@ -25,6 +24,4 @@ urlpatterns = [
 
     path('data_rights_request_sent_return', views.data_rights_request_sent_return, 
          name='data_rights_request_sent_return'),
-
-    path('pynacl_validate', pynacl_validator.validate_pynacl, name='validate_pynacl'),
 ]
